compiler/clib: Made arosc.library compile without -D__BSD_VISIBLE
git-svn-id: https://svn.aros.org/svn/aros/trunk@44596 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ compiler/clib/tmpnam.c
@@ -80,7 +80,7 @@ extern char *_mktemp(char *);
 
 ******************************************************************************/
 {
-	static u_long tmpcount;
+	static unsigned int tmpcount;
 	static char buf[L_tmpnam];
 
 	if (s == NULL)
